requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
email-validator==2.2.0
python-jose==3.5.0
bcrypt==4.3.0
pytest==8.4.1
